AndroidExtensions: move Android-specific extensions to a separate file
Also document all extension functions and fix a longstanding open-coding issue with getEncryptedPrefs
